s
s

+++ EggShell/ViewControllers/Enterprise/AllPosition/AllPosionCell.h
@@ -28,6 +28,6 @@
 //结束时间
 @property(nonatomic,strong)UILabel * endTimeLable;
 @property(nonatomic,strong)UIImageView * endTimeImageView;
-@property (strong,nonatomic)void(^isSelectBlock)(NSInteger);
+@property (strong,nonatomic)void(^isSelectBlock)(int);
 -(void)configValue:(AllPosionModel*)model;
 @end
